working but needs cleaning

+++ routes.py
@@ -1,5 +1,5 @@
 from app import app
-from flask import render_template, request, redirect
+from flask import render_template, request, redirect, flash
 from werkzeug.utils import secure_filename
 import players, downloader, comments, downloader, users
 import os
@@ -10,6 +10,10 @@ players_columns = players.get_column_names()
 app.config['MAX_CONTENT_LENGTH'] = 5000 * 5000 # this will set up max size for uploaded file
 uploads_dir = os.path.join('uploads')
 
+@app.route("/")     
+def index():        
+    user_id = users.user_id()    
+    return render_template("index.html", user_id=user_id)
 
 @app.route("/login", methods=["GET", "POST"])
 def login():
@@ -21,7 +25,8 @@ def login():
         if users.login(username, password):
             return redirect("/")
         else:
-            return render_template("error.html", error="Wrong password or username!")
+            flash("Wrong password or user doesn't exists!")
+            return redirect("/login")            
 
 @app.route("/logout")
 def logout():
@@ -37,20 +42,12 @@ def register():
         password1 = request.form["password1"]
         password2 = request.form["password2"]
         if password1 != password2:
-            return render_template("error.html", error="Passwords don't mach!")
+            flash("Passwords don't mach!")
+            return redirect("/register")
         if users.register(username, password1):
             return redirect("/")
         else:
             return render_template("error.html", error="Registration was aborted!")
-   
-
-@app.route("/")     
-def index():    
-    upload_link = ""
-    if users.user_id() != 0:
-        upload_link = "upload"
-    result = players.get_all()
-    return render_template("stats.html", players=result, players_columns=players_columns, log_link=users.log_link(), upload_link=upload_link)
 
 @app.route("/<int:id>")
 def page(id):
@@ -84,11 +81,14 @@ def send():
     players_id = request.form["id"]
     fileid = request.form["fileid"]
     if users.user_id() == 0:
-        return render_template("error.html", error="You you need to be logged in to leave a comment!")
+        flash("You you need to be logged in to leave a comment!")
+        return redirect("/files/" + fileid + "/"+ players_id)        
     if len(content) == 0:
-        return render_template("error.html", error="You are trying to send an empty comment!")
+        flash("You are trying to send an empty comment!!")
+        return redirect("/files/" + fileid + "/"+ players_id)
     if len(content) > 500:
-        return render_template("error.html", error="Your comment is too long(over 500 characters)!")
+        flash("Your comment is too long(over 500 characters)!")
+        return redirect("/files/" + fileid + "/"+ players_id)
     comments.add_comment(content, players_id, users.user_id())    
     return redirect("/files/" + fileid + "/"+ players_id)
     
@@ -100,22 +100,16 @@ def upload():
 @app.route("/uploader", methods = ["POST"])
 def upload_file():
     f = request.files["file"]    
-    if not f.filename.rsplit(".", 1)[1].lower() ==  "csv":    # here we check out that the uploaded file is csv-type
-        return render_template("error.html", error="You can only upload a csv-file!")    
-    
+    if not f.filename.rsplit(".", 1)[1].lower() ==  "csv":    # here we check out that the uploaded file is csv-type        
+        return render_template("error.html", error="You can only upload a csv-file!")        
     f.save(os.path.join(uploads_dir, secure_filename(f.filename)))
-
-    hidden = request.form.get("mycheckbox", default=False, type=bool)
-    
-    
-    print(hidden)
-    print("*****")
-    print(f.filename)
-    if downloader.user_id_and_filename_used(f.filename, users.user_id()): ## tähän muuttuja user_id kun kerran käytetään niin monesti
+    hidden = request.form.get("mycheckbox", default=False, type=bool)    
+    if downloader.user_id_and_filename_used(f.filename, users.user_id()):
         os.remove("uploads/" + f.filename)
-        return render_template("error.html", error="You have already uploaded a csv-file named '" + f.filename + "'!. If you are trying to upload a file with different content rename it.")
-    print("*****")
-    downloader.get_data(f.filename, hidden, users.user_id())
+        return render_template("error.html", error="You have already uploaded a csv-file named '" + f.filename + "'!. If you are trying to upload a file with different content rename it.")    
+    downloaded = downloader.get_data(f.filename, hidden, users.user_id())    
+    if not downloaded:        
+        return render_template("error.html", error="CSV-file have to be in the Rotowire form!")
     os.remove("uploads/" + f.filename)    
     return render_template("uploaded.html", filename=f.filename)
 
@@ -124,17 +118,14 @@ def files():
     filenames_and_ids = downloader.get_filenames_and_ids(users.user_id())
     return render_template("files.html", filenames=filenames_and_ids)
 
-@app.route("/files/<fileid>") ## allaolevissa on kyllä toisteista tietoa. kokeile saada pois
-def stats(fileid):
-    upload_link = ""
-    upload_link = users.upload_link()
+@app.route("/files/<fileid>")
+def stats(fileid):    
     filenames_and_ids = downloader.get_filenames_and_ids(users.user_id())    
     for f in filenames_and_ids:
         print(f[0])        
         if int(fileid) == f[0]:
             result = players.get_players(fileid)
-            return render_template("stats.html", players=result, players_columns=players_columns, log_link=users.log_link(), 
-            upload_link=upload_link, fileid=fileid)    
+            return render_template("stats.html", players=result, players_columns=players_columns, fileid=fileid)    
     return render_template("error.html", error="This file doesn't exists or you have no rights to view this!")
 
 @app.route("/files/<fileid>/organize/<organize>")
@@ -181,7 +172,8 @@ def liked():
     players_id = request.form["liked"]    
     fileid = request.form["fileid"]
     if users.user_id() == 0:
-        return render_template("error.html", error="You you need to be logged in to leave a like!")
+        flash("You you need to be logged in to leave a like!")
+        return redirect("/files/" + fileid + "/"+ players_id)
     users.set_liked(players_id, users.user_id())            
     return redirect("/files/" + fileid + "/"+ players_id)
